+++ resolwe_docker_images/rpkgs/packages-manual/r-packages.sh
@@ -1,0 +1,10 @@
+#!/bin/bash -e
+
+. /var/cache/build/packages-manual/common.sh
+
+# Install R packages.
+#d XXX: This is unverifiable and thus may compromise the whole image.
+# XXX: Use notary (https://github.com/ropenscilabs/notary) when ready.
+Rscript --slave --no-save --no-restore-history -e "install.packages('renv')"
+sed -e 's/^#.*$//g' -e '/^$/d' /var/cache/build/packages-r.txt | \
+  Rscript --slave --no-save --no-restore-history -e "renv::install(readLines('stdin'))"